Setup new router (#43)
